+++ api/models/__init__.py
@@ -1,3 +1,5 @@
 from .base_model import BaseModel
+from .file import File
+from .user import User
 
-__all__ = ["BaseModel"]
+__all__ = ["BaseModel", "File", "User"]